+++ R/load_input_files.R
@@ -43,6 +43,7 @@ load_input_files <- function(Image_Analyst_output, plate_metadata) {
     # check that metadata has a variable named Condition
     
     names(Input_files$metadata_df[[i]])[grepl("^(?i)(condition)$", names(Input_files$metadata_df[[i]]))] <- "Condition" # set Condition capitalization
+    names(Input_files$metadata_df[[i]])[grepl("^(?i)(ML_Training)$", names(Input_files$metadata_df[[i]]))] <- "ML_Training" # set Condition capitalization
     
     if ( !(any(grepl("^Condition$", names(Input_files$metadata_df[[i]])))) ) { # if metadata does NOT have Condition column
       enable_button_analysis()
@@ -58,14 +59,20 @@ load_input_files <- function(Image_Analyst_output, plate_metadata) {
       )
     }
     
-    # check that metadata has max 3 variables
-    if ( length(names(Input_files$metadata_df[[i]])) > 4 ) { 
+    # check that metadata has max 2 additional variables
+    add_var_names_indeces <- names(Input_files$metadata_df[[i]]) %>%
+      # remove Condition and ML_Training
+      stringr::str_detect("^Condition$|^ML_Training$|well", negate = TRUE)
+    
+    add_var_names <- names(Input_files$metadata_df[[i]])[add_var_names_indeces]
+    
+    if ( length(add_var_names) > 2 ) { 
       enable_button_analysis()
       validate(
         paste0(
-          "ERROR: Plate metadata is incorrect; too many plate templates/variables
+          "ERROR: Metadata is incorrect; too many microplate templates/variables
           
-          Ensure that each metadata file contains no more than 3 plate templates/variables (\"Condition\" + 2 optional additional variables)
+          Ensure that each metadata file contains no more than 4 microplate templates/variables (\"Condition\", \"ML_Training\" + 2 optional additional variables)
           
           incorrect metadata file: ", Input_files$metadata_name[i], "
           Plate template names/variables: ",paste(c(names(Input_files$metadata_df[[i]])), collapse=", ")